Moved Sensor into Locations folders
