+++ packages/webapp/src/containers/LocationDetails/PointDetails/SensorDetail/index.jsx
@@ -1,4 +1,4 @@
-import PureSensorDetail from '../../../components/Sensor/Detail';
+import PureSensorDetail from '../../../../components/LocationDetailLayout/Sensor/SensorDetail/index';
 
 export default function Detail({ history, user, match }) {
   return <PureSensorDetail history={history} user={user} match={match} />;
